Ensure EventSource is closed if connection fails
https://github.com/jupyterhub/binderhub/issues/317 points out that an
EventSource will repeatedly launch servers if it isn't closed properly.
Although we close the EventSource after the server launches
successfully, I noticed that the EventSource makes repeated requests if
the server fails to launch.

To test this, I opened a page that uses `nbinteract` to start a
currently banned image (`SamLau95/nbinteract-image`). In the console
logs, I saw the EventSource endlessly attempting to start a Binder
server and verified that the page made multiple requests in the Network
tab of the dev tools.

To fix this, I enforce a maximum number of failed connection attempts
and close the EventSource if it fails to connect after three tries.

To test this, I opened the same page as above and verified that the page
only makes three attempts to connect before logging an error message and
stopping.

This should help with #34, although I'm not sure if this was the reason
for the Binder overload.