Add layout flags
This commit adds the ability to mark cells with special nbinteract
comments for layouting. To mark a cell, add a comment to the
top of the cell containing one of these commands:

| Marker | Description | CSS class added |
| --------- | --------- | --------- |
| `nbi:left` | Floats cell to the left | `nbinteract-left` |
| `nbi:right` | Floats cell to the right | `nbinteract-right` |
| `nbi:hide_in` | Hides cell input | `nbinteract-hide_in` |
| `nbi:hide_out` | Hides cell output | `nbinteract-hide_out` |

The CSS classes will always be added regardless of the template, but
only the `full` template will style the classes; the `partial` and
`plain` templates leave styling up to their embedded pages.